Harden static Cypher validator ABI
